+++ package.json
@@ -8,7 +8,7 @@
     "deadcode:ci": "ts-prune | (! grep -v 'used in module') | (! grep -v 'migration') | (! grep -v 'd.ts')",
     "depcheck": "depcheck --ignores=node-fetch,@types/jest,sort-package-json",
     "depcheck:ci": "depcheck --ignores=node-fetch,@types/jest,sort-package-json | (! grep -v 'No depcheck issue') ",
-    "fetch-and-update-tempo-data": "cd src/scripts && tsx updateTempo.ts",
+    "fetch-and-update-tempo-data": "cd services && tsx updateTempo.ts",
     "lint": "eslint --ext .js,.ts .",
     "lint:ci": "eslint --ext .js,.ts . | (! grep -v 'warning') | (! grep -v 'error')",
     "sort-package:check": "sort-package-json --check",
