+++ services/run-tempo.ts
@@ -1,0 +1,5 @@
+import { updateTempoData } from "./fetchAndUpdateTempo";
+
+updateTempoData();
+
+export {};
