--- src/scripts/updateTempo.ts
@@ -1,5 +1,0 @@
-import { updateTempoData } from "./utils";
-
-updateTempoData();
-
-export {};
